moved browsing screen
